fixed all parsing valgrind issues and remove_heredoc issue

+++ sources/main.c
@@ -6,7 +6,7 @@
 /*   By: eberkowi <eberkowi@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 10:12:47 by eberkowi          #+#    #+#             */
-/*   Updated: 2024/11/08 15:29:43 by eberkowi         ###   ########.fr       */
+/*   Updated: 2024/11/19 11:25:18 by eberkowi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,8 +38,11 @@ void	remove_heredocs(t_main *main, t_tokens **tokens)
 		node = (* tokens)[i].redirects;
 		while (node != NULL)
 		{
-			if (node->type == HEREDOC)
-				unlink(node->name);
+			if (node->name)
+			{
+				if (node->type == HEREDOC && !access(node->name, F_OK))
+					unlink(node->name);
+			}
 			node = node->next;
 		}
 		i++;
@@ -69,7 +72,8 @@ int	main(int argc, char *argv[], char *envp[])
 		{
 			if (main.input)
 				free(main.input);
-			ft_free_split(&main.split_input);
+			if (main.split_input)
+				ft_free_split(&main.split_input);
 			continue;
 		}
 		execute_commandline(&main, tokens);
